requirements.txt
streamlit
pypdf
pdfplumber
pytesseract
faiss-cpu
sentence-transformers
openai
PyMuPDF
